Fix General Pension Issues
